# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EVLTE EV Charger & Accessories API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Brand {
  id: number;
  nameEn: string;
  nameRu: string;
  nameUz: string;
  slug: string;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  description?: string | null;
  order?: number;
}

export interface BrandInput {
  nameEn: string;
  nameRu: string;
  nameUz: string;
  slug: string;
  logoUrl?: string;
  description?: string;
  order?: number;
}

export interface BrandUpdate {
  nameEn?: string;
  nameRu?: string;
  nameUz?: string;
  slug?: string;
  logoUrl?: string;
  description?: string;
  order?: number;
}

export type CategoryType = typeof CategoryType[keyof typeof CategoryType];


export const CategoryType = {
  charger: 'charger',
  accessory: 'accessory',
} as const;

export interface Category {
  id: number;
  nameEn: string;
  nameRu: string;
  nameUz: string;
  type: CategoryType;
  slug?: string;
  order?: number;
}

export type CategoryInputType = typeof CategoryInputType[keyof typeof CategoryInputType];


export const CategoryInputType = {
  charger: 'charger',
  accessory: 'accessory',
} as const;

export interface CategoryInput {
  nameEn: string;
  nameRu: string;
  nameUz: string;
  type: CategoryInputType;
  slug: string;
  order?: number;
}

export type CategoryUpdateType = typeof CategoryUpdateType[keyof typeof CategoryUpdateType];


export const CategoryUpdateType = {
  charger: 'charger',
  accessory: 'accessory',
} as const;

export interface CategoryUpdate {
  nameEn?: string;
  nameRu?: string;
  nameUz?: string;
  type?: CategoryUpdateType;
  slug?: string;
  order?: number;
}

export type ProductType = typeof ProductType[keyof typeof ProductType];


export const ProductType = {
  charger: 'charger',
  accessory: 'accessory',
} as const;

export interface Product {
  id: number;
  nameEn: string;
  nameRu: string;
  nameUz: string;
  /** @nullable */
  descriptionEn?: string | null;
  /** @nullable */
  descriptionRu?: string | null;
  /** @nullable */
  descriptionUz?: string | null;
  type: ProductType;
  /** @nullable */
  powerKw?: string | null;
  price: number;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  brandId?: number | null;
  /** @nullable */
  categoryId?: number | null;
  featured?: boolean;
  inStock?: boolean;
}

export type ProductInputType = typeof ProductInputType[keyof typeof ProductInputType];


export const ProductInputType = {
  charger: 'charger',
  accessory: 'accessory',
} as const;

export interface ProductInput {
  nameEn: string;
  nameRu: string;
  nameUz: string;
  descriptionEn?: string;
  descriptionRu?: string;
  descriptionUz?: string;
  type: ProductInputType;
  powerKw?: string;
  price: number;
  imageUrl?: string;
  brandId?: number;
  categoryId?: number;
  featured?: boolean;
  inStock?: boolean;
}

export type ProductUpdateType = typeof ProductUpdateType[keyof typeof ProductUpdateType];


export const ProductUpdateType = {
  charger: 'charger',
  accessory: 'accessory',
} as const;

export interface ProductUpdate {
  nameEn?: string;
  nameRu?: string;
  nameUz?: string;
  descriptionEn?: string;
  descriptionRu?: string;
  descriptionUz?: string;
  type?: ProductUpdateType;
  powerKw?: string;
  price?: number;
  imageUrl?: string;
  brandId?: number;
  categoryId?: number;
  featured?: boolean;
  inStock?: boolean;
}

export interface BlogPost {
  id: number;
  titleEn: string;
  titleRu: string;
  titleUz: string;
  contentEn: string;
  contentRu: string;
  contentUz: string;
  /** @nullable */
  summaryEn?: string | null;
  /** @nullable */
  summaryRu?: string | null;
  /** @nullable */
  summaryUz?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  createdAt: string;
}

export interface BlogPostInput {
  titleEn: string;
  titleRu: string;
  titleUz: string;
  contentEn: string;
  contentRu: string;
  contentUz: string;
  summaryEn?: string;
  summaryRu?: string;
  summaryUz?: string;
  imageUrl?: string;
}

export interface BlogPostUpdate {
  titleEn?: string;
  titleRu?: string;
  titleUz?: string;
  contentEn?: string;
  contentRu?: string;
  contentUz?: string;
  summaryEn?: string;
  summaryRu?: string;
  summaryUz?: string;
  imageUrl?: string;
}

export interface ContactInput {
  name: string;
  email: string;
  phone?: string;
  message: string;
}

export interface ContactMessage {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  message: string;
  createdAt: string;
}

export interface SiteSettings {
  id: number;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  instagramUrl?: string | null;
  /** @nullable */
  facebookUrl?: string | null;
  /** @nullable */
  telegramUrl?: string | null;
  /** @nullable */
  youtubeUrl?: string | null;
  /** @nullable */
  addressEn?: string | null;
  /** @nullable */
  addressRu?: string | null;
  /** @nullable */
  addressUz?: string | null;
}

export interface SiteSettingsUpdate {
  phone?: string;
  email?: string;
  instagramUrl?: string;
  facebookUrl?: string;
  telegramUrl?: string;
  youtubeUrl?: string;
  addressEn?: string;
  addressRu?: string;
  addressUz?: string;
}

export type ListProductsParams = {
brandId?: number;
categoryId?: number;
powerKw?: string;
type?: ListProductsType;
featured?: boolean;
};

export type ListProductsType = typeof ListProductsType[keyof typeof ListProductsType];


export const ListProductsType = {
  charger: 'charger',
  accessory: 'accessory',
} as const;