# lib/api-zod/src/generated/types/blogPost.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EVLTE EV Charger & Accessories API
 * OpenAPI spec version: 0.1.0
 */

export interface BlogPost {
  id: number;
  titleEn: string;
  titleRu: string;
  titleUz: string;
  contentEn: string;
  contentRu: string;
  contentUz: string;
  /** @nullable */
  summaryEn?: string | null;
  /** @nullable */
  summaryRu?: string | null;
  /** @nullable */
  summaryUz?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  createdAt: string;
}